+++ test/e2e/forms/details/package-details.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Form Submission", () => {
+  test("Verify that the given details displayed on the package details page for any form type", async () => {});
+});
