+++ test/e2e/forms/spa/chip/chip-spa.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Form Submission", () => {
+  test("Create and submit a CHIP SPA", async () => {});
+});
