+++ test/e2e/forms/waiver/1915-b/comprehensive/capitated-renewal-waiver.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Form Submission", () => {
+  test("Create and submit a Capitated Renewal Waiver", async () => {});
+});
