+++ test/e2e/forms/waiver/1915-c/appendix-k-amendment.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Form Submission", () => {
+  test("Create and submit a Appendix K Amendment Waiver", async () => {});
+});
