+++ test/e2e/forms/waiver/temp-extension/temp-extension.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Form Submission", () => {
+  test("Create and submit a Temporary Extension Waiver", async () => {});
+});
