+++ test/e2e/home/user-profile.spec.ts
@@ -1,0 +1,9 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("User Profile", () => {
+  test.skip("Page header", async () => {});
+
+  test.skip("Goto IDM", async () => {});
+
+  test.skip("My Information", async () => {});
+});
