+++ test/e2e/package-actions/enable-rai-response-withdraw.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Enable RAI Response withdraw", async () => {
+  test("Perform enable RAI response withdraw for a Medicaid SPA", () => {});
+});
