+++ test/e2e/package-actions/respond-to-rai.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Enable RAI Response withdraw", () => {
+  test("Submit a Respond to RAI for a Medicaid SPA", async () => {});
+});
