+++ test/e2e/package-actions/subsequent-documents.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Enable RAI Response withdraw", () => {
+  test("Submit a subsequent documents for a Medicaid SPA", async () => {});
+});
