+++ test/e2e/package-actions/withdraw-package.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Withdraw package", () => {
+  test("Perform withdraw package action for a Medicaid SPA", async () => {});
+});
