+++ test/e2e/package-actions/withdraw-rai-response.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Withdraw RAI Response", async () => {
+  test("Perform withdraw RAI response for a Medicaid SPA", () => {});
+});
