+++ test/e2e/package-actions/withdraw/disable-rai-response-withdraw.spec.ts
@@ -1,0 +1,5 @@
+import { test } from "@playwright/test";
+
+test.describe.skip("Disable RAI Response withdraw", () => {
+  test("Perform disable RAI response withdraw for a Medicaid SPA", async () => {});
+});
