+++ test/e2e/playwright.config.ts
@@ -20,7 +20,7 @@ console.log(`Playwright configured to run against ${baseURL}`);
 export default defineConfig({
   // need to find a reasonable timeout less than 30s
   // timeout: 10_000,
-  testDir: ".",
+  testDir: "../../test",
   /* Run tests in files in parallel */
   fullyParallel: true,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
