+++ test/e2e/tsconfig.json
@@ -21,5 +21,5 @@
       "@/*": ["src/*"]
     }
   },
-  "include": ["."]
+  "include": [".", "playwright.config.ts"]
 }
