+++ test/pages/index.ts
@@ -1,0 +1,3 @@
+// export * from "./login.page";
+// export * from "./home.page";
+// export * from "./faq.page";
