+++ test/pages/sample.page.ts
@@ -1,0 +1,9 @@
+// import { type Page } from "@playwright/test";
+
+// export class SamplePage {
+//   readonly page: Page;
+
+//   constructor(page: Page) {
+//     this.page = page;
+//   }
+// }
